--- torchvision/prototype/models/detection/ssdlite.py
@@ -1,124 +1,0 @@
-import warnings
-from functools import partial
-from typing import Any, Callable, Optional
-
-from torch import nn
-from torchvision.models._api import WeightsEnum, Weights
-from torchvision.models._meta import _COCO_CATEGORIES
-from torchvision.models._utils import handle_legacy_interface, _ovewrite_value_param
-from torchvision.models.detection.ssdlite import (
-    _mobilenet_extractor,
-    _normal_init,
-    _validate_trainable_layers,
-    DefaultBoxGenerator,
-    det_utils,
-    SSD,
-    SSDLiteHead,
-)
-from torchvision.models.mobilenetv3 import MobileNet_V3_Large_Weights, mobilenet_v3_large
-from torchvision.transforms import ObjectDetectionEval, InterpolationMode
-
-
-__all__ = [
-    "SSDLite320_MobileNet_V3_Large_Weights",
-    "ssdlite320_mobilenet_v3_large",
-]
-
-
-class SSDLite320_MobileNet_V3_Large_Weights(WeightsEnum):
-    COCO_V1 = Weights(
-        url="https://download.pytorch.org/models/ssdlite320_mobilenet_v3_large_coco-a79551df.pth",
-        transforms=ObjectDetectionEval,
-        meta={
-            "task": "image_object_detection",
-            "architecture": "SSDLite",
-            "publication_year": 2018,
-            "num_params": 3440060,
-            "size": (320, 320),
-            "categories": _COCO_CATEGORIES,
-            "interpolation": InterpolationMode.BILINEAR,
-            "recipe": "https://github.com/pytorch/vision/tree/main/references/detection#ssdlite320-mobilenetv3-large",
-            "map": 21.3,
-        },
-    )
-    DEFAULT = COCO_V1
-
-
-@handle_legacy_interface(
-    weights=("pretrained", SSDLite320_MobileNet_V3_Large_Weights.COCO_V1),
-    weights_backbone=("pretrained_backbone", MobileNet_V3_Large_Weights.IMAGENET1K_V1),
-)
-def ssdlite320_mobilenet_v3_large(
-    *,
-    weights: Optional[SSDLite320_MobileNet_V3_Large_Weights] = None,
-    progress: bool = True,
-    num_classes: Optional[int] = None,
-    weights_backbone: Optional[MobileNet_V3_Large_Weights] = None,
-    trainable_backbone_layers: Optional[int] = None,
-    norm_layer: Optional[Callable[..., nn.Module]] = None,
-    **kwargs: Any,
-) -> SSD:
-    weights = SSDLite320_MobileNet_V3_Large_Weights.verify(weights)
-    weights_backbone = MobileNet_V3_Large_Weights.verify(weights_backbone)
-
-    if "size" in kwargs:
-        warnings.warn("The size of the model is already fixed; ignoring the parameter.")
-
-    if weights is not None:
-        weights_backbone = None
-        num_classes = _ovewrite_value_param(num_classes, len(weights.meta["categories"]))
-    elif num_classes is None:
-        num_classes = 91
-
-    trainable_backbone_layers = _validate_trainable_layers(
-        weights is not None or weights_backbone is not None, trainable_backbone_layers, 6, 6
-    )
-
-    # Enable reduced tail if no pretrained backbone is selected. See Table 6 of MobileNetV3 paper.
-    reduce_tail = weights_backbone is None
-
-    if norm_layer is None:
-        norm_layer = partial(nn.BatchNorm2d, eps=0.001, momentum=0.03)
-
-    backbone = mobilenet_v3_large(
-        weights=weights_backbone, progress=progress, norm_layer=norm_layer, reduced_tail=reduce_tail, **kwargs
-    )
-    if weights_backbone is None:
-        # Change the default initialization scheme if not pretrained
-        _normal_init(backbone)
-    backbone = _mobilenet_extractor(
-        backbone,
-        trainable_backbone_layers,
-        norm_layer,
-    )
-
-    size = (320, 320)
-    anchor_generator = DefaultBoxGenerator([[2, 3] for _ in range(6)], min_ratio=0.2, max_ratio=0.95)
-    out_channels = det_utils.retrieve_out_channels(backbone, size)
-    num_anchors = anchor_generator.num_anchors_per_location()
-    assert len(out_channels) == len(anchor_generator.aspect_ratios)
-
-    defaults = {
-        "score_thresh": 0.001,
-        "nms_thresh": 0.55,
-        "detections_per_img": 300,
-        "topk_candidates": 300,
-        # Rescale the input in a way compatible to the backbone:
-        # The following mean/std rescale the data from [0, 1] to [-1, -1]
-        "image_mean": [0.5, 0.5, 0.5],
-        "image_std": [0.5, 0.5, 0.5],
-    }
-    kwargs: Any = {**defaults, **kwargs}
-    model = SSD(
-        backbone,
-        anchor_generator,
-        size,
-        num_classes,
-        head=SSDLiteHead(out_channels, num_anchors, num_classes, norm_layer),
-        **kwargs,
-    )
-
-    if weights is not None:
-        model.load_state_dict(weights.get_state_dict(progress=progress))
-
-    return model
